Revert "update google play to 5 (#1762)"
This reverts commit f78dafb39722efacb7b49e7ac94ad523813932cc.
