+++ .github/workflows/ci-android.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install base modules
         run: yarn; yarn build
       - name: Kotlin Linter
-        uses: vroy/gha-kotlin-linter@v4
+        uses: vroy/gha-kotlin-linter@v3
       - name: Grant execute permission for gradlew
         run: chmod +x android/gradlew
       - name: Build with Library with Gradle and check lints
