+++ src/types/index.ts
@@ -60,7 +60,7 @@ export interface ProductPurchase {
   quantityIOS?: number;
   originalTransactionDateIOS?: string;
   originalTransactionIdentifierIOS?: string;
-  //Android V4
+  //Android
   dataAndroid?: string;
   signatureAndroid?: string;
   autoRenewingAndroid?: boolean;
@@ -70,29 +70,6 @@ export interface ProductPurchase {
   developerPayloadAndroid?: string;
   obfuscatedAccountIdAndroid?: string;
   obfuscatedProfileIdAndroid?: string;
-  //Android V5
-  title?: string;
-  description?: string;
-  productType?: string;
-  name?: string;
-  oneTimePurchaseOfferDetails?: {
-    priceCurrencyCode?: string;
-    formattedPrice?: string;
-    priceAmountMicros?: string;
-  }[];
-  subscriptionOfferDetails?: {
-    offerToken?: string[];
-    pricingPhases: {
-      pricingPhaseList: {
-        formattedPrice?: string;
-        priceCurrencyCode?: string;
-        billingPeriod?: string;
-        billingCycleCount?: number;
-        priceAmountMicros?: string;
-        recurrenceMode?: number;
-      };
-    };
-  }[];
   //Amazon
   userIdAmazon?: string;
   userMarketplaceAmazon?: string;
